requirements.txt
torch~=2.2.1
Flask~=3.0.2
flasgger~=0.9.7.1
diffusers~=0.27.2
requests==2.31.0
flask-restful==0.3.9
transformers==4.39.1



Pillow~=9.4.0
numpy~=1.26.2
torchvision~=0.17.1
